New search using search.json in browser to search
search.json is generated on the server and contains the text of all pages referenced in the navigation and additionally the blog posts.

There are new javascrip files: js/jquery-2.2.4.min.js and js/search.js. Where
jquery-2.2.4.min.js is a copy of the official jquery-2.2.4. and search.js is
own code to enable the search in the browser.