Fix mixed-up usage of in- and egress for policies
In the man-pages for the firewalld.policies the usage of "ingress" and
"egress" are mixed-up, so "ingress" is mentioned for the "OUTPUT" chain
and "egress" is mentioned for the "INPUT" chain.
This commit fixes this so "ingress" is mentioned for the "INPUT" chain
and "egress" is mentioned for the "OUTPUT" chain.